requirements.txt
python-dotenv
pandas
numpy
google-cloud-storage
certifi
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart


-e .